Added Electron build + fixed cursor on Electron.
Added build types 8+9 (Electron BASIC + HIBASIC).

Removed writes to the 6845 registers (as they're not present on the
Electron) in favour of VDU 23:

Cursor now flashes in insert mode and is steady in overwrite mode and
does not crash when screen is scrolled.

Minor changes to leave the cursor flashing in command mode and when
exiting the editor.