Command codes now use functional names not keys
This is to prepare the code to support the Electron, where the
available keys are totally different: this will allow the command names
to remain the same in mainpart.s65 but just be changed in cmds.s65.

Also, tidied the list of commands to organise it by position and added
the cursor + COPY key codes.

These changes shouldn't result in any actual change to the output
binary.